Move GeneOnotologies to shared components.
